Properly Removed Old Config Files
Now that the Config XML rework is complete, we won't be needing these anymore. Strangely bittersweet to do lmao, especially for EX_Characters.txt. It's been in the repo for more than 2 years now lol, and represents the very first code contribution I ever made to the PM community! Since then I've gotten to do so much more, from the further configuration options in for this menu builder, to the Kirby Hat stuff in 2021, to the Super Sawndz revamp last year, and the recent work on the ASM/Gecko Translator. Just, idk, it feels really good to be able to look back at where we started and *feel* the growth lol. I feel like we've hit a very real stride, and I'm really excited to see where we'll go and what we'll do next lmao. But yeah lmfao, just real happy lol, super grateful for the past couple years I've gotten to serve this community, and here's to another two!