Reworked Jumpsquat Modifier, Added Random Option
Because the game only ever expects to receive integer values as jumpsquat frame lengths, I needed to ensure that values were always properly rounded to zero; easiest way to do that was simply to apply our changes before we convert to float in the first place.

Additionally, implemented a random jumpsquat length option, for when normal delinquency isn't enough.