Added Distance Option to Jumpsquat Modifier
This calculates a character's new jumpsquat length based on the absolute distance between the given Modifier Value and the character's natural jumpsquat. So for example, with a value of 8, a character with 3 jumpsquat would end up at ABS(8 - 3) = 5, and a character with 5 would end up with ABS(8 - 5) = 3. Additionally, a value of 3 would have a character with 5 end up at ABS(3- 5) = 2.