Initial Rework of PPC Block Parser Logic
Moves a lot of the block analysis logic into a struct to bundle everything up more neatly. Though, I'm not sure if this is actually the best way to do this or not, since we only care about the state of that struct while we're parsing a PPC block in the first place, so... Idk, I'll figure that part out later, but this does work for now thankfully.

Additionally, data embeds have received some rather significant changes: they're applied all at once at the end of the translation process, and there's now a mechanism in-place for disabling data embed detection in blocks which use certain combinations of instructions that make it impossible to guarantee that certain regions of code are actually unreachable. Really, we should also be blacklisting codes that also use mtctr and bctr, since technically you can mflr then mtctr to skirt the existing blacklist, but I feel like losing the ability to detect embeds at all in codes which call Brawl functions is just too much of a loss for how little we gain. And even if we incorrectly identify an embed, it should still compile the same, it'll just be labeled wrong, so I mean... fair trade, I feel like.