Lines Can Now Be Made Unselectable
Added a field to the Line class which lets you mark them as unselectable! Additionally, so that unselectable lines could use a distinct color from other lines, the menu's line color system has been reworked into an expandable table at the end of the Code Menu Header. This allows for the trivial introduction and removal of additional line colors, and should open the door for the implementation of additional line types going forwards, should the need for them arise.