Rewrote and Commented Options XML Parsing Code
I wrote the code for reading in default options and things from the Options XML a *long* time ago, before I was as familiar with pugixml as I am now, and it really shows in how I approached pulling different elements from the XML itself. So, I've gone through and rewritten everything in a much more "sane" way lmao, should be both easier to understand (both by virtue of its approach, and the new comments) and more efficient.