Reworked Character List Collision Handling Code
Previously, it was possible to use entries in the characterList block of the Config XML to change the ID associated with an existing character by specifying an entry with the same name, but a different ID.
I wanted to kinda replicate that ability in reverse (make it possible to change the *name* of an existing entry by specifying an entry with the same ID), which ultimately required a bit of a rewrite to how the logic actually worked.
The new implementation supports this, and additionally produces more detailed log messages both when collisions occur, *and* when an entry can't be processed because neither its Name or ID were unique.