Added ".orig" Backup Functionality
When the program goes to backup and overwrite a file, it'll first check if a ".orig" version of that file exists. If one doesn't, it'll make a copy of the file to be overwritten with ".orig" appended to the back, which should act as a longer-term backup of the file as it was the first time this program was run (or at least, the first time a version with this functionality was run).

Additionally, reimplemented copyFile() function using the std::filesystem copy function.