Made Debug Settings Configurable via XML
Additionally, if the program detects that a "debugOptions" block is present, but is missing a defined debug option, it'll automatically add it to the Config XML for you! So if you just need to get all the debug options, you can make an empty debugOptions node and run the program, it'll auto-populate it with all the options!