Fixed Crash on Initial File Backup
I didn't realize originally that this function threw exceptions if one of the files doesn't exist (which seems silly to me, but on me for not reading lmfao). Fixed, in any case: we now use the normal "==" operator, which checks lexicographically rather than seeing if the paths resolve the same, but that's good enough for our needs.