HEAP_CACHE_TABLE Now Separately Stores Heap IDs
I'd originally assumed that heaps couldn't actually move during gameplay, so I wrote this system assuming that I'd only need to cache the values once. This actually *isn't* true; Heaps *can* move, so I need to be able to repeatedly update the cache as values change, which wasn't possible with the way I originally set things up.

This commit changes things around a bit, introducing a (word-aligned) array containing the Heap IDs for each cached Heap Address, which we'll be using to rebuild the cache as the appropriate values as needed.