Initialized MEM2 Variable Block in Header to 0xCC
Hadn't occurred to me that code behavior for some of those values might depend on them starting at a value of 0xCC. Before this change, the SavePreviousReplay line for example caused a crash if you attempted to use it before you'd started a match, because the game evaluated IS_IN_GAME_FLAG to false on startup instead of true, since its value was 0x00 on startup instead of 0xCC, which caused the game to attempt to save a non-existent replay lmao. This fixes that problem, the MEM2 variables are once again in a space akin to their native environment.