Implemented Saturation and Luminence Modifiers
Each color is now represented by a triple of floats, with the first being the Hue, the second being a Saturation modifier, and the third being a Luminence modifier! These modifiers span from 0.0f to 2.0f, with values below 1.0f scaling the associated property down to 0.0f, and values above 1.0f scaling it up to its maximum value!