PSCC Now Respects UserData Disable Masks
The code now checks the UserData block of the CLR0 for a Disable Mask corresponding to the current Material, and will selectively enable or disable itself based on whether the bit corresponding to any given material-target is set.

Note: Material-targets are considered enabled by default! If there is no UserData to draw the Mask from, or if there is no Mask corresponding to a given Material, the code will be activated for all associated material-targets!