Flipped Roles of User-Def and CLR0 Float Trip Vals
The second and third values specifically are what was flipped; previously, the idea was that CLR0s defined a Hue modifier, then absolute Saturation and Luminance; while the Code Menu had defined absolute Hue, and Saturation and Luminance modifiers. This resulted in some inflexibility in defining exact colors, so now the user defines all 3 of the absolute values (ie. they give the color they want in HSL format). The CLR0s' jobs now is to take those absolute colors and transform them as necessary to get the desired end result.

Reworked color float constants in the mean time to work with this flip, though again, they're getting externalized very soon.