Heavy Optimizations to Constant Overrides Code
The old version of this code did a *ton* of redundant register setup between reads and writes; this does away with that. By keeping track of the top half of the read and write address registers, we're able to only set those values as necessary, using just the immediate in the LWZ and STW instructions to move the data we want! Winds up cutting that portion of the code more than in half, down to 31 instructions from 70!