Improvements to PSCC Config Logging
Added a final color list summary, moved the code for adding colors to colorTable out into its own function, and cleaned up the way the function was written a little.

Also, specific color objects are now initialized with a specific hue value before they're read in from the Config XML; if they still have that value when when we go to commit their changes, they'll be skipped and won't emit an error message!