Initial Line Hiding Impl. + Added Sticky Line Flag
Implements utilities for marking lines as "hidden", which will force the line to the end of its parent page, and force its default color to Black to make it invisible when not selected or altered. Combining this with the "unselectable" flag will make the line unselectable as well, effectively removing it from the code menu in any player-noticeable sense.

This system is fully Options XML compatible as well, allowing these designations to be freely added to and removed from lines at will, without the need to recompile the builder source!