Fixes for Missed PPC Data Embeds
Catches some situations where data embeds weren't being caught correctly, particularly in situations where the instruction that initiates an embed is also the destination of another branch. Also fixes situations where the branch that initiates an embed doesn't land at the end of the embed.

Lastly, fixed an issue in the Map File parser where addresses that landed on the first instruction of a function wouldn't register as part of that function.