vP+ Config No Longer Disables Char List Additions
Just some minor adjustments to enable this and ensure the logging is clean with it on. Cuz after thinking about it, it isn't *technically* necessary to lock this in the first place. If for some reason someone's base P+ build needs to use it, more power to them. Not adding the stuff for it to the stock "Config.xml" since most people won't be needing it, but the functionality is enabled at least.