Continued PSCC Flags Impl.
Decided ultimately to scrap the ability to differentiate the Hue flags on the direction of the modifier. The delineation I was trying to draw felt somewhat arbitrary, it complicated the actual logic I'd have to do to implement it quite a lot, and I don't think was really useful in the same way being able to individual control upwards and downwards Sat and Lum mods was.
I've also elected to cut the flags short down to a single char, since the callback system ultimately I think will be a lot more useful than room for 8 more flag functionalities.