Internalized PSCC Callback Table + RGB Strobe
I'd previously implemented these as a separate ASM file, but figure I ought to make them properly internal just to ensure they're always present. As such, removed the validation check for the table pointer, as by definition if the code for checking the pointer is present the code for writing the pointer is as well.