Finished Core Addon Functionality
Implemented the core functionality for writing addon lines' INDEX values out to the CMNU, as well as for preparing .aliases for their LOC values in a dedicated file, and appending .include statements for each addon's .asm file's in the main CodeMenu.asm!

That marks the feature completion of the addon system! What's left now is to hook up adequate logging, shuffle around some string constants, and comment up the code itself.