Initial Switch Table Impl + Toy ExecAction Reimpl
These were implemented mostly as a curiosity, but ultimately I don't think I'm gonna be using it lol. My hope was that a jump table would help to optimize down certain bits of code (ExecuteAction being a prime candidate, in my original plans), but ultimately the space necessary for just the boilerplate stuff like embedding the table and moving things to and from CTR/LR bloat things to a point where this'll never be more efficient outside of situations with *a loooot* of cases lol. Still, fun exercise though; and the switch table implementation *does* work, so I'm leaving the code here for future reference (in case I ever wanna come back to it).