Reworked Code Menu Activation "Off" Setting
Replaces the old functionality of the "Off" option (resetting the menu on exit) with the ability to just completely disable Code Menu access via controller inputs, which I expect will be much more useful for most people. Also gets rid of the only other instance of the "ResetPage()" function, which should make optimizing it into a subroutine a bit simpler.

Also, disabled mnemonic output for `crandc`, since GCTRM currently is compiling it correctly.