Added Empty Base List and Char Removal Support
Implements functionality to allow users to start from an empty list and add characters from there, and additionally to remove characters from the list using entries with a special flag.

Additionally, since it's now possible to configure the menu such that the character list is *empty*, I've implemented a check which ensures that the character switcher is entirely removed if the list ends up completely empty.