Added Addon WorkingMem and Count Caps
Just a precaution to prevent anyone from accidentally creating mega-big menu files using the addon system, either intentionally or unintentionally. Limits are pretty sensible I think as well, we allow a maximum allocation of 1kb per addon, and a maximum of 256 addons total, so the resulting menu file *can't* end up with any more than 256kb of Addon data in it.

Note that staying beneath this limit in no way guarantees that the menu file will be able to load without conflict into the game; this is only here as a precaution to protect people from potential dangerous addons.